pyproject.toml
[tool.poetry]
name = "act"
version = "0.0.0"
description = ""
authors = ["Josh Wycuff <Josh.Wycuff@gmail.com>"]
readme = "README.md"
packages = [{include = "act"}]

[tool.poetry.dependencies]
python = "^3.9"
PyYAML = "^6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
act = "act.main:main"
